+++ covador/__init__.py
@@ -2,7 +2,7 @@ from .utils import wrap_in, dpass
 from .vdecorator import ValidationDecorator
 from .types import *
 
-version = '0.9.22'
+version = '0.9.23'
 
 schema = make_schema(item_getter)
 list_schema = make_schema(list_item_getter)